Fix the check whether some optional variables are defined
This issue came up while testing this role in the bootstraping instance
where no user_groups and sudoers_group was defined.

While doing so figured out that the creation of the users fails if no
(additional) users groups where defined. By default, the primary user
group should be the same as the username.