Use current (and stable) implementation of role user
The main reason for doing so was the way the sudoers file was created.
In the previous implementation the given sudoers_groups list was used to
create /etc/sudoers.d/{{ item }} files with the following content:

  %{{ item }} ALL=NOPASSWD: ALL

which means that all members of the group have passwordless sudo rights.

In the current stable implementation the sudoers file is named
/etc/sudoers.d/{{ groupname }} and has all group members (given by
user_names) listed inside:

  username1 ALL=NOPASSWD: ALL
  username2 ALL=NOPASSWD: ALL

There is one difference to the current role user implementation (taken
from synpro/ansible.git rev b17dc4b164d41f236dfb77ea3dbba795cb2ab3b2):

This role does *not* deploy the grml-config to be independent from the
role base.