Do not use /bin/zsh as default
When using the role user in the bootstraping process we can not assume
that zsh is installed. When changing the user shell if the given shell
is not installed a login is not possible.

In fact, I am not even sure if should set a user shell. It might be
better to keep the default or detect the installed shells and try a list
with possible fallbacks, like: /bin/zsh -> /bin/bash -> /bin/sh