Do not set /bin/bash if user_shell is not given
Initially the role did always set the user shell to /bin/zsh. But this
caused problems on systems where zsh was not installed. So the variable
user_shell was introduced and /bin/bash seemed like a good fall back if
the variable was unset. It turned out, that if a role like grml-config
changes the shell to anything other than /bin/bash this role reverts
this change back to /bin/bash (if user_shell was not given). So the best
way seems to only set the default shell if this is explicilty given.