Remove user_shell from defaults
Forgot to remove user_shell from defaults. But after doing so it turned
out that user_shell is undefined which is not allow. (Maybe we can set
an empty string to keep the shell?).
Anyway, added a seperate task to set the shell if user_shell is defined
and while at reorder create_user.yml to be used in with_items loop. It
should be noted that import_tasks does not support loops, so I had to
got with include_tasks.