Update ansible molecule test
Replaced the molecule test foled with an up-to-date molecule:

  % molecule --version
  molecule 3.1.5
      ansible:2.9.6 python:3.7
      delegated:3.1.5 from molecule
      docker:0.2.4 from molecule_docker

  % molecule init scenario -d docker -r ansible-role-user

Created a simple molecule test to run again Debian/buster.

In the meanwhile created a simple Github Workflow 'Ansible Molecule',
so removed .travis.yml.

ansible-lint also found that tasks/create_authorized_keys.yml had errors.
It turned out that the file was unused so removed it.