Migrate Ansible Molecule from docker to podman
The local system was updated from Debian/bullseye to Debian/bookworm.
Podman works well, so we switched Ansible Molecule from docker to
podman.

We now also use the Ansible version in Debian/bookworm
(v2.14.3), so we removed Ansible from the virtual environment.

Updated "Getting Started" link to ansible molecule.

robertdebock.bootstrap 7.0.2 seems to be included in the podman image,
so removed it from the dependencies.