Revert "Rework ansible role jkirk.users + implement user_groups"
This reverts commit 57abbd1900dc3c6193c06e32137d7659592ef60f.

This also reverts "molecule converge update" where we tested
"user_groups", see: 4abc44c and 0b01edd.

In aca6d78 we changed the group ownership of the authorized_keys file
from the username group to the new primary group. While doing so we
implicitly fixed an issue where the file was deployed multiple time.

So this only partly reverts aca6d78 and we take over that fix and we now
deploy the authorized_keys while we create the users and do not iterate
through all the users multiple times.

Issue: #5